Fix the HTML errors in... wrong way?
When Markdown generates HTML errors and I'm trying to fix HTML errors
in non-Turing complete designing language, all I can do is chain of
replacing and removing. It doesn't even have regexpes!

Also, it seems that this language is sort of whitespace sensitive in
the wrong way. Anything other than one space would cause syntax errors.
This includes linebreaks - this is why chain of replaces is on one
line.